Observe templates folder for new templates
